backend/pyproject.toml
[project]
name = "collectioner-backend"
version = "1.0.0"
description = ""
authors = [{ name = "Petr", email = "petr.ishimov@outlook.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
  
  "django (>=4.2,<5.0)",
  "djangorestframework (>=3.14,<3.16)",
  "djangorestframework-simplejwt (>=5.3,<6.0)",
  "django-cors-headers (>=4.0,<5.0)",
  "drf-spectacular (>=0.27,<0.28)",
  "drf-spectacular-sidecar (>=2024.4,<2026.0)",
  "pillow (>=11.3.0,<12.0.0)",
  "bs4 (>=0.0.2,<0.0.3)",
  "httpx (>=0.28.1,<0.29.0)",
  "celery[redis] (>=5.5.3,<6.0.0)",
  "django-celery-beat (>=2.8.1,<3.0.0)",
  "psycopg[binary] (>=3.2.12,<4.0.0)",
  "dj-database-url (>=3.0.1,<4.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
  "django-filter (==24.3)",
]

[tool.poetry.group.ci.dependencies]
pytest = ">=9.0.0,<10.0.0"
pytest-django = ">=4.11.1,<5.0.0"
pytest-cov = ">=7.0.0,<8.0.0"
factory-boy = ">=3.3.3,<4.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
isort = "^5.13"
pylint = "^3.0"
pylint-django = "^2.5"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["apps", "core", "config", "server"]